requirements.txt
# requirements.txt para lector_gdrive

google-api-python-client
google-auth
google-auth-oauthlib
